Added search capability. #3
User can start to type in the input box and a list of breeds ill appear. They can choose from a breed in the list or continue typing. When they hit search, an exact match search is performed on the data set. Any pet that matches the requested breed is shown. Searching for an empty value will reveal all the animals again.